legal-chatbot-frontend/src/final_scripts/requirements.txt
langchain_openai
langsmith
langgraph
numexpr
langchain-community
tiktoken
langchainhub
unstructured
python-dateutil
unstructured[pdf]
nltk
pathway
pathway[all]
pathway[xpack-llm]
openai
llama-index
llama-index-retrievers-pathway
langchain_voyageai
langfuse
anthropic
langchain_anthropic
langchain_cohere
scikit-learn
transformers
python-dotenv
transformers
pymongo
